Add SetFormat method to wxDatePickerGeneric
